Ensure a non-__main__ module when testing, and ensure the module is properly imported when unserializing. Also fix dynamic functions calling a builtin.
